Rename to HTTP Adapter
