+++ catowser/BrowserNetworking/Extensions/AlamofireHTTPAdaptee.swift
@@ -1,5 +1,5 @@
 //
-//  AlamofireExtensions.swift
+//  AlamofireHTTPAdaptee.swift
 //  HttpKit
 //
 //  Created by Andrei Ermoshin on 11/29/21.
@@ -13,7 +13,7 @@ import ReactiveSwift
 import Combine
 #endif
 
-final class AFNetworkingBackend<RType: ResponseType, SType: ServerDescription>: HTTPNetworkingBackend {
+final class AlamofireHTTPAdaptee<RType: ResponseType, SType: ServerDescription>: HTTPAdapter {
     typealias TYPE = RType
     typealias SRV = SType
     
@@ -93,53 +93,6 @@ final class AFNetworkingBackend<RType: ResponseType, SType: ServerDescription>: 
         if case .waitsForCombinePromise = handlerType {
             let promiseWrapper: HttpKit.CombinePromiseWrapper<RType, SType> = .init(promise, endpoint)
             handlerType = .combine(promiseWrapper)
-        }
-    }
-}
-
-final class AFNetworkingVoidBackend: HTTPNetworkingBackendVoid {
-    var wrapperHandler: ((Result<Void, HttpKit.HttpError>) -> Void)
-    
-    var handlerType: ResponseVoidHandlingApi
-    
-    init(_ handlerType: ResponseVoidHandlingApi) {
-        self.handlerType = handlerType
-        wrapperHandler = handlerType.wrapperHandler
-    }
-    
-    func performVoidRequest(_ request: URLRequest, sucessCodes: [Int]) {
-        let dataRequest: DataRequest = AF.request(request)
-        dataRequest
-            .validate(statusCode: sucessCodes)
-            .response { [weak self] (defaultResponse) in
-                let result: Result<Void, HttpKit.HttpError>
-                if let error = defaultResponse.error {
-                    let localError = HttpKit.HttpError.httpFailure(error: error)
-                    result = .failure(localError)
-                } else {
-                    let value: Void = ()
-                    result = .success(value)
-                }
-                self?.wrapperHandler(result)
-        }
-        if case let .rxObserver(_, lifetime) = handlerType {
-            lifetime.observeEnded({
-                dataRequest.cancel()
-            })
-        } else if case let .combine(_) = handlerType {
-            // https://github.com/kyzmitch/Cotton/issues/14
-        }
-    }
-    
-    func transferToRxState(_ observer: Signal<Void, HttpKit.HttpError>.Observer, _ lifetime: Lifetime) {
-        if case .waitsForRxObserver = handlerType {
-            handlerType = .rxObserver(observer, lifetime)
-        }
-    }
-    
-    func transferToCombineState(_ promise: @escaping Future<Void, HttpKit.HttpError>.Promise) {
-        if case .waitsForCombinePromise = handlerType {
-            handlerType = .combine(promise)
         }
     }
 }
