+++ catowser/BrowserNetworking/HttpClient+Alamofire.swift
@@ -10,33 +10,33 @@ import HttpKit
 import ReactiveSwift
 
 extension HttpKit.Client {
-    public func makePublicRequest<T, B: HTTPNetworkingBackend>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                               networkingBackend: B) where B.TYPE == T, B.SRV == Server {
+    public func makePublicRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+                                                     networkingBackend: B) where B.TYPE == T, B.SRV == Server {
         makeCleanRequest(for: endpoint, withAccessToken: nil, networkingBackend: networkingBackend)
     }
     
-    public func makeAuthorizedRequest<T, B: HTTPNetworkingBackend>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                                   withAccessToken accessToken: String,
-                                                                   networkingBackend: B) where B.TYPE == T, B.SRV == Server {
+    public func makeAuthorizedRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+                                                         withAccessToken accessToken: String,
+                                                         networkingBackend: B) where B.TYPE == T, B.SRV == Server {
         makeCleanRequest(for: endpoint, withAccessToken: accessToken, networkingBackend: networkingBackend)
     }
     
-    public func rxMakePublicRequest<T, B: HTTPNetworkingBackend>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                                 networkingBackend: B) -> SignalProducer<T, HttpKit.HttpError> where B.TYPE == T, B.SRV == Server {
+    public func rxMakePublicRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+                                                       networkingBackend: B) -> SignalProducer<T, HttpKit.HttpError> where B.TYPE == T, B.SRV == Server {
         let producer = rxMakeRequest(for: endpoint, withAccessToken: nil, networkingBackend: networkingBackend)
         return producer
     }
     
-    public func rxMakeAuthorizedRequest<T, B: HTTPNetworkingBackend>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                                     withAccessToken accessToken: String,
-                                                                     networkingBackend: B) -> SignalProducer<T, HttpKit.HttpError> where B.TYPE == T, B.SRV == Server {
+    public func rxMakeAuthorizedRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+                                                           withAccessToken accessToken: String,
+                                                           networkingBackend: B) -> SignalProducer<T, HttpKit.HttpError> where B.TYPE == T, B.SRV == Server {
         let producer = rxMakeRequest(for: endpoint, withAccessToken: accessToken, networkingBackend: networkingBackend)
         return producer
     }
     
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
-    public func cMakePublicRequest<T, B: HTTPNetworkingBackend>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                                networkingBackend: B) -> ResponseFuture<T> where B.TYPE == T, B.SRV == Server {
+    public func cMakePublicRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+                                                      networkingBackend: B) -> ResponseFuture<T> where B.TYPE == T, B.SRV == Server {
         let future = cMakeRequest(for: endpoint,
                                      withAccessToken: nil,
                                      networkingBackend: networkingBackend)
@@ -44,9 +44,9 @@ extension HttpKit.Client {
     }
     
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
-    public func cMakeAuthorizedRequest<T, B: HTTPNetworkingBackend>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                                    withAccessToken accessToken: String,
-                                                                    networkingBackend: B) -> ResponseFuture<T> where B.TYPE == T, B.SRV == Server {
+    public func cMakeAuthorizedRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+                                                          withAccessToken accessToken: String,
+                                                          networkingBackend: B) -> ResponseFuture<T> where B.TYPE == T, B.SRV == Server {
         let future = cMakeRequest(for: endpoint,
                                      withAccessToken: accessToken,
                                      networkingBackend: networkingBackend)
