+++ catowser/HttpKit/HttpClient+Combine.swift
@@ -16,9 +16,9 @@ extension HttpKit.Client {
     public typealias ResponseFuture<T> = Publishers.HandleEvents<Deferred<Future<T, HttpKit.HttpError>>>
     
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
-    public func cMakeRequest<T, B: HTTPNetworkingBackend>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                          withAccessToken accessToken: String?,
-                                                          networkingBackend: B) -> ResponseFuture<T> where B.TYPE == T, B.SRV == Server {
+    public func cMakeRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+                                                withAccessToken accessToken: String?,
+                                                networkingBackend: B) -> ResponseFuture<T> where B.TYPE == T, B.SRV == Server {
         return Combine.Deferred {
             let subject: Future<T, HttpKit.HttpError> = .init { [weak self] (promise) in
                 guard let self = self else {
