+++ catowser/HttpKit/HttpClient+RxSwift.swift
@@ -10,9 +10,9 @@ import Foundation
 import ReactiveSwift
 
 extension HttpKit.Client {
-    public func rxMakeRequest<T, B: HTTPNetworkingBackend>(for endpoint: HttpKit.Endpoint<T, Server>,
-                                                           withAccessToken accessToken: String?,
-                                                           networkingBackend: B) -> SignalProducer<T, HttpKit.HttpError> where B.TYPE == T, B.SRV == Server {
+    public func rxMakeRequest<T, B: HTTPAdapter>(for endpoint: HttpKit.Endpoint<T, Server>,
+                                                 withAccessToken accessToken: String?,
+                                                 networkingBackend: B) -> SignalProducer<T, HttpKit.HttpError> where B.TYPE == T, B.SRV == Server {
         let producer: SignalProducer<T, HttpKit.HttpError> = .init { [weak self] (observer, lifetime) in
             guard let self = self else {
                 observer.send(error: .zombieSelf)
